Port Prometheus components to TS
