+++ package.json
@@ -27,6 +27,7 @@
   },
   "homepage": "https://github.com/matrix-org/matrix-appservice-bridge#readme",
   "dependencies": {
+    "@types/express": "^4.17.6",
     "bluebird": "^2.9.34",
     "extend": "^3.0.0",
     "is-my-json-valid": "^2.19.0",
@@ -36,7 +37,7 @@
     "nedb": "^1.1.3",
     "nopt": "^3.0.3",
     "p-queue": "^6.3.0",
-    "prom-client": "^11.1.1",
+    "prom-client": "^12.0.0",
     "request": "^2.61.0"
   },
   "devDependencies": {
